Improve gitlab plugin.
Moving from using 1 gitlab function with "action" argument to
individual functions for each action. This provides a better API
because in the old situation the args you would have to provide
changed depending on the action. Therefor it was not easy to spot
which arguments you had to specify. If we move to individual functions
for each action then it becomes more clear, because it is in the
functio defenition.

Adding URL encoding for URLs that take a file path. This allows us to
specify to take action on files that are in a directory (files that
are not in the root of the repository).